Bug: 339298 Description: Merge sun_merge_branch_20060523 changes into trunk. Fix Description: I had to change re_encode_request to use the new ber types.  Other than that, these merges are basically the sasl, ipv6, and ber types code already committed to the sun merge branch.

+++ directory/c-sdk/ldap/libraries/libldap/compare.c
@@ -137,7 +137,7 @@ ldap_compare_ext( LDAP *ld, const char *dn, const char *attr,
 	}
 
 	if ( ber_printf( ber, "{it{s{so}}", *msgidp, LDAP_REQ_COMPARE, dn,
-	    attr, bvalue->bv_val, (int)bvalue->bv_len /* XXX lossy cast */ )
+	    attr, bvalue->bv_val, bvalue->bv_len )
 	    == -1 ) {
 		lderr = LDAP_ENCODING_ERROR;
 		LDAP_SET_LDERRNO( ld, lderr, NULL, NULL );
